+++ pycommon_server/celery_common.py
@@ -1,3 +1,4 @@
+from typing import Tuple
 import datetime
 import json
 import logging
@@ -10,6 +11,7 @@ from celery import Celery
 from celery import result as celery_results
 from celery.task import control
 from flask_restplus import Resource, fields, Namespace
+from redis import Redis
 
 _STATUS_ENDPOINT = "status"
 _RESULT_ENDPOINT = "result"
@@ -229,6 +231,51 @@ def _snake_case(name: str) -> str:
         raise ValueError(f"{name} should be Camel Case and should not contain any _")
     s1 = re.sub("(.)([A-Z][a-z]+)", r"\1_\2", name)
     return re.sub("([a-z0-9])([A-Z])", r"\1_\2", s1).lower()
+
+
+def redis_health_details(redis_url: str, namespace: str) -> Tuple[str, dict]:
+    redis = Redis.from_url(redis_url)
+    try:
+        redis.ping()
+    except Exception as e:
+        return (
+            "fail",
+            {
+                "redis:ping": {
+                    "componentType": "component",
+                    "status": "fail",
+                    "time": datetime.datetime.utcnow().isoformat(),
+                    "output": str(e),
+                }
+            },
+        )
+
+    keys = redis.keys(namespace)
+
+    if not keys:
+        return (
+            "fail",
+            {
+                "redis:ping": {
+                    "componentType": "component",
+                    "status": "fail",
+                    "time": datetime.datetime.utcnow().isoformat(),
+                    "output": f"missing namespace: {namespace}",
+                }
+            },
+        )
+
+    return (
+        "pass",
+        {
+            "redis:ping": {
+                "componentType": "component",
+                "observedValue": keys,
+                "status": "pass",
+                "time": datetime.datetime.utcnow().isoformat(),
+            }
+        },
+    )
 
 
 def health_details(config: dict):